+roadsplit

+++ python3/jpegoptimizer3.py
@@ -9,30 +9,44 @@
 
 import os
 import sys
-import cv2
+# import cv2
 import numpy as np
 from tqdm import tqdm
 from PIL import Image
+from PIL import ImageOps
+# import piexif
 
 
-print ("*** JPEG optimizer ***")
-try:
-    ordnerpfad = sys.argv[1]
-except:
-    ordnerpfad = r"D:\Mapillary\DCIM"
-    pass
+def optimize_file(file_path):
+    image_org = Image.open(file_path)
+    try:
+        exif_data = image_org.info['exif']
+        image_rgb = ImageOps.autocontrast(image_org)
+        image_rgb.save(file_path, optimize=True, quality=75, exif=exif_data)
+    except Exception as error:
+        print(error)
+        image_org.close()
+        return False
+    return True
 
-for subdirz, dirz, filez in os.walk(ordnerpfad):
-    print("\nOptimizing:", subdirz)
-    for f in tqdm(filez):
-        file_path = subdirz + os.sep + f
-        if file_path.lower().endswith('.jpg'):
-            image_org = Image.open(file_path)
-            try:
-                exifdata = image_org.info['exif']
-                image_cv2 = cv2.normalize(np.array(image_org), None, 0, 255, cv2.NORM_MINMAX)
-                image_rgb = Image.fromarray(image_cv2)
-                image_rgb.save(file_path, optimize=True, quality=75, exif=exifdata)
-            except:
-                image_org.close()
-                os.remove(file_path)
+
+#
+#   Main
+#
+if __name__ == "__main__":
+    print ("*** JPEG optimizer ***")
+    try:
+        ordnerpfad = sys.argv[1]
+    except:
+        ordnerpfad = r"D:\Mapillary\DCIM"
+        pass
+    # Loop over JPG files
+    for subdirz, dirz, filez in os.walk(ordnerpfad):
+        print("\nOptimizing:", subdirz)
+        for f in tqdm(filez):
+            file_path = subdirz + os.sep + f
+            if file_path.lower().endswith('.jpg'):
+                success = optimize_file(file_path)
+                if not(success):
+                    os.remove(file_path)
+
